Update project structure.
This changes the cabal file layout slightly.  We use a nix derivation of
network-arbitrary-0.6.0.0.  We also use brittany to format the source
code.  The last update is an updated gitignore file.

This is just to keep the project structure up to date with other new
projects I've been toying with.